patch: applied/libventi-rpc-discon
Handle RPCs in libventi when not connected to a Venti server.

This keeps the behavior more consistant with the old libventi.
The old libventi handle this case properly in the function vtRPC
(/sys/src/liboventi/client.c).

It has been reviewed and merged in plan9port.

Reference: /n/sources/patch/applied/libventi-rpc-discon
Date: Wed Jan 18 10:10:36 CET 2012
Signed-off-by: djc@9grid.fr