patch: maybe/myipaddr-thread
myipaddr was not thread safe.  there are two free-loaders with this
patch, a bit of formatting for ip(2), which may need to be redone
in a better way, and the expansion of ip statstics to uvlongs to prevent
overflows.

Notes:
Mon Mar 17 18:18:19 EDT 2014 geoff
    mostly dubious.

Reference: /n/sources/patch/maybe/myipaddr-thread
Date: Sun Feb 19 14:47:57 CET 2012
Signed-off-by: quanstro@quanstro.net
Reviewed-by: geoff