patch: applied/fossil-archive-corruption
When _cacheLocal loads the label of a block from the disk and not its
contents (if called with mode = OOverWrite), mark the cache entry
iostate as BioLabel.  It used to be marked BioClean, which causes
problems when doRemoveLink calls cacheLocalData with OOverWrite and
then leaves the data invalid, where it can be found by a concurrent
archThread and written to venti.

Reference: /n/sources/patch/applied/fossil-archive-corruption
Date: Thu Mar 22 17:26:04 CET 2012
Signed-off-by: miller@hamnavoe.com