patch: applied/venti-bloom-writedonechan
Fix writedonechan size in Venti Bloom.

It fixes a crash on Venti Bloom, when
running the Plan 9 64-bit kernel.

The writedonechan Channel was incorrectly
allocated with a size of void* instead
unsigned long. This is a problem on Plan 9
because the type long is always 32-bit on
both 32-bit and 64-bit systems.

This change has been reviewed and merged
in plan9port.

Reference: /n/sources/patch/applied/venti-bloom-writedonechan
Date: Wed May  2 07:57:14 CES 2012
Signed-off-by: djc@9grid.fr