patch: applied/vacfs-srv-orclose
vacfs: fix create srv with ORCLOSE

In the current code, the srv file is removed
just after the main thread exits, while the
srv thread is still running, which is not
the expected behavior.
We moved the srv creation just before the
procrfork, in order that the srv file will
not be removed until the srv thread exits.

This change has been reviewed and merged
in plan9port.

Reference: /n/sources/patch/applied/vacfs-srv-orclose
Date: Fri Aug  3 21:21:48 CES 2012
Signed-off-by: djc@9grid.fr