patch: maybe/u9fs-chroot
This patch address two long-standing issues in u9fs:

First, use of chroot causes complications when a root is specified; chroot requires /etc/^(passwd group) and friends be copied to the new root.  This is uneccessary, and is obviated by use of a new function named rootpath.

Second, files created with a default user are owned by the user of the process rather than the default user.  This is addressed by issuing a chown after create to reinstate proper permission.

Reference: /n/sources/patch/maybe/u9fs-chroot
Date: Mon Aug 20 06:52:41 CES 2012
Signed-off-by: sstallion@gmail.com