patch: applied/make-big-macro-suicide
proper fix for suicide in make when building svn.

This was caused by some vast macro expansions writing off then end
of some fixed arrays.

I increased the size of the arrays and added a check on the end of the
array so make dies with a diagnositc rather than a random suicide

-Steve

Reference: /n/sources/patch/applied/make-big-macro-suicide
Date: Tue Jan 22 23:09:39 CET 2013
Signed-off-by: steve@quintile.net