patch: maybe/no-sender-ml
upas/ml adds the sender of mails to a list to the recipient list.
Though it checks out for dup addresses, most of the times the sender
address is not exactly like the one on the list, which means dup mail
is received. Also, the sender might think he's subscribed because he gets
a copy of the mail.

This comments out the code adding the sender.

Notes:
Wed Aug 7 06:58:18 CES 2013 bootes

Reference: /n/sources/patch/maybe/no-sender-ml
Date: Mon Mar  4 11:59:45 CET 2013
Signed-off-by: nemo@lsub.org
Reviewed-by: geoff