patch: maybe/mkinstall-mkdir
It is a burden to create /*/bin/disk or /*/bin/dir_for_your_new_tool,
and it is easier if mk install creates the BIN directory.
A minor problem is that if you make a mistake you are going to create
a directory that should not be there; but in that case you can remove it.

Reference: /n/sources/patch/maybe/mkinstall-mkdir
Date: Thu Mar 14 19:38:24 CET 2013
Signed-off-by: nemo@lsub.org