patch: applied/vncv-curtrail
When using vncv on a terminal with software cursor (vesa, rpi) the
mouse cursor leaves a trail.  This seem to be caused by the fact that
vncv loads picture updates with loadimage(2) directly to screen.
Loading to an offscreen image followed by a draw(2) to screen removes
the artifact.

Reference: /n/sources/patch/applied/vncv-curtrail
Date: Tue May  7 11:01:08 CES 2013
Signed-off-by: yarikos@gmail.com