patch: applied/import-m
This is similar to mycroftiv's import-z, but uses the -m option (already in use for the same
function at another site), which gives the mnemonic "mount-only", instead of -z.
More important, it doesn't require an argument that's then ignored, which avoids
having to document an ignored argument (which would be unique, I think).
Instead, if -m is given, you must not provide the tree (ie, if you do, it's diagnosed).
I'd mentioned all this on 9fans, but perhaps the suggestions were missed in the long
conversations. There are no patents on this option.

Reference: /n/sources/patch/applied/import-m
Date: Tue May  7 12:58:51 CES 2013
Signed-off-by: charles.forsyth@gmail.com