patch: ktrace-link-reg
This patch addresses a long-standing bug and usability issue.

For architectures that do not rely exclusively on a link register
(such as ARM), ktrace would complain if a link register was not
provided as an argument.  The diagnostic was superfluous at best, and
inaccurate as zero is a valid link register value.

Finally, an acid line was printed to ease usability when analyzing
traces.

Reference: /n/sources/patch/ktrace-link-reg
Date: Thu Jul  4 11:36:26 CES 2013
Signed-off-by: sstallion@gmail.com