patch: maybe/rdbfs-retry-speed-broken
This patch addresses a few long-standing bugs and usability issues.

Retries have been corrected (previously if rdbfs had to retry, it
would always fail).

Setting the baudrate for the underlying serial console has also been
made optional.  Forcing the baudrate to 9600 causes issues on systems
that default to a different value.

Status was changed such that rdbfs processes now show up in the Broken
state with a meaningful user and name.  A note file was also added to
send a serial break on a kill note.

Reference: /n/sources/patch/maybe/rdbfs-retry-speed-broken
Date: Thu Jul  4 22:48:04 CES 2013
Signed-off-by: sstallion@gmail.com