patch: applied/arm-bkpt-cond
In armv6 and armv7, breakpoint instruction is conditionally executed, so use ALWAYS
condition in the instruction to ensure the breakpoint will be activated whatever the
condition code is in the PSR.  Unfortunately on kirkwood (and possibly other armv5?)
this changes the exception type from Undefined Instruction to Prefetch Abort, so trap.c
must be changed to allow for both possibilities.

After applying this patch, rebuild db and acid, and /arm/9plug if applicable.

Reference: /n/sources/patch/applied/arm-bkpt-cond
Date: Fri Aug  2 23:03:49 CES 2013
Signed-off-by: miller@hamnavoe.com