patch: applied/acme-warp
From rsc by way of quanstro:

"from russ.  if a window is deleted with Del in the tag, and a new
window is moved in to replace it, then warp the mouse to Del.  this
facilliates the common case of deleting a bunch of windows at the same
time.  the theory is that the random pointer location in the newly
moved window wouldn't be useful anyway."

Reference: /n/sources/patch/applied/acme-warp
Date: Mon Aug 12 09:29:33 CES 2013
Signed-off-by: sstallion@gmail.com