patch: maybe/cpu-noauth
This patch adds support for a -n flag to cpu(1), which may be used
with the -R or -O flags to disable authentication for incoming
requests.  This option must be enabled explicitly as it can lead to
some rather undesirable behavior in an untrusted environment.

Notes:
Sun Feb 23 01:27:40 EST 2014 geoff
    this seems really dubious and probably not worth distributing.

Reference: /n/sources/patch/maybe/cpu-noauth
Date: Sun Sep 22 09:12:58 CES 2013
Signed-off-by: sstallion@gmail.com
Reviewed-by: geoff