patch: maybe/telnet-notkbd
This patch corrects a couple of minor problems with telnet when using
the notkbd option (-n).

The first issue was to disable changing raw mode when a pipe was being
used.  This would result in raw mode being enabled on the calling
console.

The second issue was a deadlock caused by the net child being handed a
terminal note.  This would cause the parent to spin waiting for the
child to notify.  This was changed to a waitpid loop such that the
main thread can shutdown gracefully if needed.

Reference: /n/sources/patch/maybe/telnet-notkbd
Date: Thu Dec  5 06:09:07 CET 2013
Signed-off-by: sstallion@gmail.com