patch: sorry/quanstro-auth
the purpose of this patch is to make out-of-the box plan 9
authenticate against coraid, my own site and 9atom.  since there
isn't a way to discover authentication, this data otherwise needs
to be added by hand.  this can be difficult during an install.  these
are really three independent requests, but it seemed easier to
do a single patch.

Reference: /n/sources/patch/sorry/quanstro-auth
Date: Sun Jan  5 01:27:37 CET 2014
Signed-off-by: quanstro@quanstro.net