patch: srvaname
allow the attach name to be specified in srv(4).
this is helpful when writing rules for 9fs that need
an attachpoint.  the standard way of doing this

case aggle
	srv $nflag -q tcp!aggle.example.org aggle &&
		mount $nflag /srv/aggle /n/aggle aggle

suffers the problem that if /srv/aggle becomes stale, it is
not retried.  with this change this rule can be rewritten
like so

case aggle
	srv $nflag -q tcp!aggle.example.org aggle /n/aggle aggle

and it is now correct, even if /srv/aggle becomes stale.

Reference: /n/sources/patch/srvaname
Date: Tue Apr  1 04:05:12 CES 2014
Signed-off-by: quanstro@quanstro.net