patch: broke-add-new-a-flag
Add a new '-a' flag to broke(1); makes it print all Broken processes.

'broke -a' saves a lot of time when you are trying to track down busted
processes on a server.  I don't like running day-to-day as the host owner,
or someone in sys, so broke tends not to show me things I am interested in
as the sysadmin.  'broke -a' shows things that are borked on a server, and gives the
heads-up that one of the hardware or users need a tune up.

Reference: /n/sources/patch/broke-add-new-a-flag
Date: Tue May 27 01:36:19 CES 2014
Signed-off-by: lyndon@orthanc.ca